+++ server/routes/index.js
@@ -1,21 +1,7 @@
-var express = require("express");
-var router = express.Router();
+const express = require("express");
+const router = express.Router();
 const services = require("../services/render");
 const controller = require("../controller/controller");
-const mysql = require("mysql");
-const sha256 = require("sha256");
-const url = require("url");
-const jwt = require("jsonwebtoken");
-var querystring = require("querystring");
-const db = mysql.createConnection({
-  host: "158.247.222.53",
-  user: "dev",
-  password: "13975",
-  database: "schoolmate",
-});
-db.connect();
-
-const bodyParser = require("body-parser");
 
 /* GET home page. */
 router.get("/", services.homeRoutes);
